extracts discount
