+++ lib/object_wrappers.dart
@@ -7,3 +7,4 @@ export 'offering_wrapper.dart';
 export 'offerings_wrapper.dart';
 export 'errors.dart';
 export 'payment_discount.dart';
+export 'discount.dart';
